requirements.txt
fastmcp==0.2.0
anthropic==0.40.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
httpx==0.27.0
pydantic==2.7.0